+++ src/common/utils/deployTips.ts
@@ -4,35 +4,62 @@ import * as format from "string-template";
 import stripAnsi = require("strip-ansi");
 import { getAllTips } from "./deployTipsList";
 import { deployErrorsToMarkdown, testFailuresToMarkdown } from "../gitProvider/utilsMarkdown";
+import { askOpenApi, isOpenApiAvailable } from "./openaiUtils";
+import { ChatCompletion } from "openai/resources";
+import { uxLog } from ".";
 
 let logRes = null;
 let errorsAndTips = [];
+let alreadyProcessedErrors = [];
 const firstYellowChar = c.yellow("*")[0];
 
 // Checks for deploy tips in a log string
 // returns formatted and completed error log
 export async function analyzeDeployErrorLogs(log: string, includeInLog = true, options: any): Promise<any> {
   errorsAndTips = []; // reset
+  alreadyProcessedErrors = [] // reset
   logRes = returnErrorLines(log).join("\n"); // reset
   const tips: any = [];
   for (const tipDefinition of getAllTips()) {
-    if (matchesTip(tipDefinition, includeInLog)) {
+    if ((await matchesTip(tipDefinition, includeInLog))) {
       tips.push(tipDefinition);
     }
   }
   // Add default error messages for errors without tips
   const logResLines = [];
   const updatedLogLines = returnErrorLines(logRes);
-  updatedLogLines.forEach((logLine, index) => {
+  let index = 0;
+  for (const logLine of updatedLogLines) {
     logResLines.push(logLine.trim());
-    if (logLine.trim().startsWith("Error") && !(updatedLogLines[index + 1] && !updatedLogLines[index + 1].trim().startsWith("Error"))) {
-      logResLines.push(c.yellow("No sfdx-hardis tip to solve this error. Try google ?"));
-      logResLines.push(c.yellow(""));
-      errorsAndTips.push({
-        error: { message: stripAnsi(logLine.trim()) },
-      });
+    if (logLine.includes("Error (1): Deploy failed.")) {
+      index++;
+      continue;
     }
-  });
+    if (logLine.trim().startsWith("Error") && !(updatedLogLines[index + 1] && !updatedLogLines[index + 1].trim().startsWith("Error"))) {
+      const aiTip = await findAiTip(logLine.trim());
+      // Complete with AI if possible
+      if (aiTip && aiTip.choices.length > 0) {
+        logResLines.push(c.yellow((`[AI] ${aiTip.model} suggested the following tip ${c.bold(c.bgRed('(can be good or stupid, this is AI !)'))}:`)));
+        logResLines.push(c.magenta(c.italic(aiTip.choices[0].message.content)));
+        logResLines.push(c.yellow(""));
+        errorsAndTips.push({
+          error: { message: stripAnsi(logLine.trim()) },
+          tipFromAi: aiTip.choices[0].message.content
+        });
+      }
+      else {
+        // No tip found, give the user an AI prompt
+        logResLines.push(c.yellow("No sfdx-hardis tip to solve this error. You can try the following prompt:"));
+        logResLines.push(c.yellow(buildPrompt(logLine.trim())));
+        logResLines.push(c.yellow(""));
+        errorsAndTips.push({
+          error: { message: stripAnsi(logLine.trim()) },
+        });
+      }
+
+    }
+    index++;
+  }
 
   // Extract failed test classes
   const failedTests = [];
@@ -66,14 +93,13 @@ export async function analyzeDeployErrorLogs(log: string, includeInLog = true, o
       }
     }
   }
-
   updatePullRequestResult(errorsAndTips, failedTests, options);
   return { tips, errorsAndTips, failedTests, errLog: logResLines.join("\n") };
 }
 
 // Checks if the error string or regex is found in the log
 // Adds the fix tip under the line if includeInLog is true
-function matchesTip(tipDefinition: any, includeInLog = true): boolean | any {
+async function matchesTip(tipDefinition: any, includeInLog = true): Promise<boolean | any> {
   const newLogLines = [];
   // string matching
   if (
@@ -86,8 +112,10 @@ function matchesTip(tipDefinition: any, includeInLog = true): boolean | any {
       const logLines = returnErrorLines(logRes);
       for (const line of logLines) {
         newLogLines.push(line);
+        let found = false;
         for (const expressionString of tipDefinition.expressionString) {
           if (line.includes(expressionString)) {
+            found = true;
             newLogLines.push(c.yellow(c.italic("Tip for " + tipDefinition.label + ":")));
             newLogLines.push(...tipDefinition.tip.split(/\r?\n/).map((str: string) => c.yellow(str)));
             newLogLines.push(c.yellow(" "));
@@ -99,6 +127,18 @@ function matchesTip(tipDefinition: any, includeInLog = true): boolean | any {
                 message: tipDefinition.tip,
               },
             });
+          }
+        }
+        if (found) {
+          const aiTip = await findAiTip(line.trim());
+          // Complete with AI if possible
+          if (aiTip && aiTip.choices.length > 0) {
+            newLogLines.push(c.yellow((`[AI] ${aiTip.model} suggested the following tip ${c.bold(c.bgRed('(can be good or stupid, this is AI !)'))}:`)));
+            newLogLines.push(c.magenta(c.italic(aiTip.choices[0].message.content)));
+            newLogLines.push(c.yellow(""));
+            const lastErrorAndTip = errorsAndTips[errorsAndTips.length - 1];
+            lastErrorAndTip.tipFromAi = aiTip.choices[0].message.content;
+            errorsAndTips[errorsAndTips.length - 1] = lastErrorAndTip;
           }
         }
       }
@@ -119,10 +159,12 @@ function matchesTip(tipDefinition: any, includeInLog = true): boolean | any {
       const logLines = returnErrorLines(logRes);
       for (const line of logLines) {
         newLogLines.push(line);
+        let found = false;
         for (const expressionRegex of tipDefinition.expressionRegex) {
           expressionRegex.lastIndex = 0; // reset regex last index to be able to reuse it
           const matches = [...line.matchAll(expressionRegex)];
           for (const m of matches) {
+            found = true ;
             const replacements = m.map((str: string) => c.bold(str.trim().replace(/'/gm, "")));
             const replacementsMarkdown = m.map((str: string) => `**${str.trim().replace(/'/gm, "")}**`);
             newLogLines.push(c.yellow(c.italic(format(tipDefinition.label, replacements))));
@@ -137,6 +179,18 @@ function matchesTip(tipDefinition: any, includeInLog = true): boolean | any {
                 message: stripAnsi(format(tipDefinition.tip, replacementsMarkdown).replace(/\*\*.\*\*/gm, ".")),
               },
             });
+          }
+          if (found) {
+            const aiTip = await findAiTip(line.trim());
+            // Complete with AI if possible
+            if (aiTip && aiTip.choices.length > 0) {
+              newLogLines.push(c.yellow((`[AI] ${aiTip.model} suggested the following tip ${c.bold(c.bgRed('(can be good or stupid, this is AI !)'))}:`)));
+              newLogLines.push(c.magenta(c.italic(aiTip.choices[0].message.content)));
+              newLogLines.push(c.yellow(""));
+              const lastErrorAndTip = errorsAndTips[errorsAndTips.length - 1];
+              lastErrorAndTip.tipFromAi = aiTip.choices[0].message.content;
+              errorsAndTips[errorsAndTips.length - 1] = lastErrorAndTip;
+            }
           }
         }
       }
@@ -170,4 +224,29 @@ async function updatePullRequestResult(errorsAndTips: Array<any>, failedTests: A
     prData.status = "invalid";
   }
   globalThis.pullRequestData = Object.assign(globalThis.pullRequestData || {}, prData);
+}
+
+async function findAiTip(errorLine: any): Promise<ChatCompletion | null> {
+  if (alreadyProcessedErrors.includes(errorLine)) {
+    return null ;
+  }
+  alreadyProcessedErrors.push(errorLine);
+  if (isOpenApiAvailable()) {
+    const prompt = buildPrompt(errorLine);
+    try {
+      const aiResponse = await askOpenApi(prompt);
+      return aiResponse;
+    } catch (e) {
+      uxLog(this, c.yellow("[AI] Error while calling OpenAI: " + e.message));
+    }
+  }
+  return null;
+}
+
+function buildPrompt(errorLine: string) {
+  const prompt = `How to solve Salesforce deployment error "${errorLine}" ? \n` +
+    "- Please answer using sfdx source format, not metadata format. \n" +
+    "- Please provide XML example if applicable. \n" +
+    "- Please skip the part of the response about retrieving or deploying the changes with Salesforce CLI.";
+  return prompt;
 }