+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- Integrate with OpenAI ChatGPT to find solutions to deployment issues
+
 ## [4.40.2] 2024-06-18
 
 - hardis:org:diagnose:audittrail: Define new not suspect actions